+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased](https://github.com/edtr-io/edtr-io/compare/0.3.1..HEAD)
+## [0.3.2](https://github.com/edtr-io/edtr-io/compare/0.3.1..0.3.2) - March 14, 2019
 
 ### Breaking Changes
 
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Image plugin
+- Added Image plugin
 - Added `editable` prop to `Editor` for switching between edit and render mode
 - Added optional `onPaste` callback to Plugins, which is called by the text plugin on paste event
 - Added `InlineOverlay` Component for plugin controls